fix(73): use preventUnauthorizedInterceptor request config param, update axios

+++ packages/web/src/services/http/http.service.ts
@@ -1,17 +1,23 @@
-import axios from 'axios';
+import axios, { AxiosRequestConfig } from 'axios';
 import { api } from 'config';
 
-export const http = axios.create({
+export interface Config extends AxiosRequestConfig {
+  preventUnauthorizedInterceptor?: boolean;
+}
+
+const config: Config = {
   baseURL: api,
-});
+};
+
+export const http = axios.create(config);
 
 http.interceptors.request.use(
-  config => {
+  request => {
     const token = localStorage.getItem('authToken');
     if (token) {
-      config.headers.authorization = token; // eslint-disable-line no-param-reassign
+      request.headers.authorization = token; // eslint-disable-line no-param-reassign
     }
-    return config;
+    return request;
   },
   error => Promise.reject(error),
 );
@@ -19,11 +25,9 @@ http.interceptors.request.use(
 http.interceptors.response.use(
   response => response,
   error => {
-    if (error.response.status === 401) {
-      if (window.location.pathname !== '/signin') {
-        localStorage.clear();
-        window.location.pathname = '/signin';
-      }
+    if (error.response.status === 401 && !error.config.preventUnauthorizedInterceptor) {
+      localStorage.clear();
+      window.location.pathname = '/signin';
     }
     return Promise.reject(error);
   },
